Renamed the regression_test package to regressiontest
