+++ regression-tests/src/main/java/de/jplag/regressiontest/constants/TestDirectoryConstants.java
@@ -1,4 +1,4 @@
-package de.jplag.regression_test.constants;
+package de.jplag.regressiontest.constants;
 
 import java.nio.file.Path;
