+++ regression-tests/src/main/java/de/jplag/regressiontest/helper/DeltaSummaryStatistics.java
@@ -1,4 +1,4 @@
-package de.jplag.regression_test.helper;
+package de.jplag.regressiontest.helper;
 
 import java.util.DoubleSummaryStatistics;
