+++ regression-tests/src/main/java/de/jplag/regressiontest/helper/FileHelper.java
@@ -1,4 +1,4 @@
-package de.jplag.regression_test.helper;
+package de.jplag.regressiontest.helper;
 
 import java.io.File;
 import java.io.FileOutputStream;
