+++ regression-tests/src/main/java/de/jplag/regressiontest/model/DataSetFormat.java
@@ -1,4 +1,4 @@
-package de.jplag.regression_test.model;
+package de.jplag.regressiontest.model;
 
 import java.io.File;
 import java.io.IOException;
