+++ regression-tests/src/main/java/de/jplag/regressiontest/model/ExpectedResult.java
@@ -1,4 +1,4 @@
-package de.jplag.regression_test.model;
+package de.jplag.regressiontest.model;
 
 import de.jplag.JPlagComparison;
 import de.jplag.options.SimilarityMetric;
