+++ regression-tests/src/main/java/de/jplag/regressiontest/model/GoldStandard.java
@@ -1,4 +1,4 @@
-package de.jplag.regression_test.model;
+package de.jplag.regressiontest.model;
 
 import java.util.Collection;
 import java.util.DoubleSummaryStatistics;
